clean after merge
