+++ starknet_py/tests/e2e/fixtures/account_clients.py
@@ -1,6 +1,7 @@
 # pylint: disable=redefined-outer-name
 
 import sys
+from dataclasses import dataclass
 from typing import Tuple, List, Dict
 
 import pytest
@@ -217,17 +218,33 @@ def account_client(request) -> AccountClient:
     return request.getfixturevalue(request.param)
 
 
+@dataclass
+class AccountToBeDeployedDetailsFactory:
+
+    class_hash: int
+    fee_contract: Contract
+
+    async def get(self) -> AccountToBeDeployedDetails:
+        return await get_deploy_account_details(
+            class_hash=self.class_hash, fee_contract=self.fee_contract
+        )
+
+
 @pytest_asyncio.fixture(scope="module")
-async def details_of_account_to_be_deployed(
+async def deploy_account_details_factory(
     account_with_validate_deploy_class_hash: int,
     fee_contract: Contract,
-) -> AccountToBeDeployedDetails:
+) -> AccountToBeDeployedDetailsFactory:
     """
-    Returns address, key_pair, salt and class_hash of the account with validate deploy.
+    Returns AccountToBeDeployedDetailsFactory.
+
+    The Factory's get() method returns: address, key_pair, salt
+    and class_hash of the account with validate deploy.
     Prefunds the address with enough tokens to allow for deployment.
     """
-    return await get_deploy_account_details(
-        class_hash=account_with_validate_deploy_class_hash, fee_contract=fee_contract
+    return AccountToBeDeployedDetailsFactory(
+        class_hash=account_with_validate_deploy_class_hash,
+        fee_contract=fee_contract,
     )
